requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
PyJWT>=2.8.0
httpx>=0.25.0
python-multipart>=0.0.6
